refactor: move tests/ into frontend/tests/
The Playwright e2e/integration suite is frontend-specific and its specs
already imported from frontend/src, so nest it under frontend/ to declutter
the repo root.

- git mv tests/ -> frontend/tests/
- update cross-dir imports (drop redundant frontend/ segment)
- bump tests tsconfig extends ../ -> ../../
- update package.json scripts (playwright config path, oxfmt targets)
- update CI seed.sql path and docs/development.md references
