+++ frontend/tests/lexicons.spec.ts
