+++ frontend/tests/navigation.spec.ts
