+++ frontend/tests/observation-detail.spec.ts
