+++ package.json
@@ -13,13 +13,13 @@
     "generate-rust-types": "./scripts/generate-rust-types.sh",
     "setup": "./scripts/setup.sh",
     "dev": "vite -c frontend/vite.config.ts",
-    "fmt": "oxfmt frontend/src tests scripts",
-    "fmt:check": "oxfmt --check frontend/src tests scripts",
+    "fmt": "oxfmt frontend/src frontend/tests scripts",
+    "fmt:check": "oxfmt --check frontend/src frontend/tests scripts",
     "lint": "oxlint frontend/src",
     "storybook": "storybook dev -p 6006",
-    "test": "playwright test --config=tests/playwright.config.ts",
-    "test:e2e": "playwright test --config=tests/playwright.config.ts --project=e2e",
-    "test:integration": "playwright test --config=tests/playwright.config.ts --project=integration",
+    "test": "playwright test --config=frontend/tests/playwright.config.ts",
+    "test:e2e": "playwright test --config=frontend/tests/playwright.config.ts --project=e2e",
+    "test:integration": "playwright test --config=frontend/tests/playwright.config.ts --project=integration",
     "test:unit": "vitest run --config frontend/vitest.config.ts"
   },
   "dependencies": {
